+++ tests/test_fields.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 from __future__ import with_statement
 import datetime
 import os
@@ -198,9 +199,6 @@ class FieldTest(unittest.TestCase):
 
         link.url = 'http://www.google.com:8080'
         link.validate()
-
-        link.url = u'http://президент.рф'
-        self.assertTrue(link.validate())
 
     def test_int_validation(self):
         """Ensure that invalid values cannot be assigned to int fields.