+++ docs/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 
 Changes in 0.7.X
 ================
-- Updated URLField - can handle unicode and custom validator (MongoEngine/mongoengine#136)
+- Updated URLField - can have a custom validator (MongoEngine/mongoengine#136)
 - Allow Django AuthenticationBackends to work with Django user (hmarr/mongoengine#573)
 - Fixed reload issue with ReferenceField where dbref=False (MongoEngine/mongoengine#138)